Refactor (1)
